+++ .formatter.exs
@@ -1,0 +1,5 @@
+[
+  import_deps: [:ecto],
+  inputs: ["*.{ex,exs}", "{config,lib,test}/**/*.{ex,exs}"],
+  subdirectories: []
+]