+++ lib/ecto_job/migrations.ex
@@ -65,12 +65,14 @@ defmodule EctoJob.Migrations do
     ## Options
 
     * `:prefix` - the prefix (aka Postgresql schema) to create the table in.
+    * `:version` - the major version of the EctoJob library used to generate the table
     * `:timestamps` - A keyword list of options passed to the `Ecto.Migration.timestamps/1` function.
     """
     def up(name, opts \\ []) do
       opts = [{:primary_key, false} | opts]
       prefix = Keyword.get(opts, :prefix)
       timestamp_opts = Keyword.get(opts, :timestamps, [])
+      version = Keyword.get(opts, :version, 2)
 
       _ =
         create table(name, opts) do
@@ -87,7 +89,24 @@ defmodule EctoJob.Migrations do
           add(:max_attempts, :integer, null: false, default: 5)
           add(:params, :map, null: false)
           add(:notify, :string)
+
+          if version >= 3 do
+            add(:priority, :integer, null: false, default: 0)
+          end
+
           timestamps(timestamp_opts)
+        end
+
+      _ =
+        case version do
+          1 ->
+            nil
+
+          2 ->
+            create(index(name, [:schedule, :id]))
+
+          3 ->
+            create(index(name, [:priority, :schedule, :id]))
         end
 
       execute("""
@@ -109,6 +128,37 @@ defmodule EctoJob.Migrations do
       prefix = Keyword.get(opts, :prefix)
       execute("DROP TRIGGER tr_notify_inserted_#{name} ON #{Helpers.qualify(name, prefix)}")
       execute("DROP TABLE #{Helpers.qualify(name, prefix)}")
+    end
+  end
+
+  defmodule UpdateJobTable do
+    @moduledoc """
+    Defines an update migration to an especific version of Ecto Job.
+    This migration can be run multiple times with different values to update multiple queues.
+    """
+
+    import Ecto.Migration
+
+    @doc """
+    Upgrade the job queue table with the given ecto job version and name.
+    """
+    def up(3, name) do
+      alter table(name) do
+        add(:priority, :integer, null: false, default: 0)
+      end
+
+      create(index(name, [:priority, :schedule, :id]))
+    end
+
+    @doc """
+    Rollback updates from job queue table with the given ecto job version and name.
+    """
+    def down(3, name) do
+      _ = drop(index(name, [:priority, :schedule, :id]))
+
+      alter table(name) do
+        remove(:priority)
+      end
     end
   end
 end