+++ test/support/migrations/20170817092638_add_job_table.exs
@@ -2,9 +2,11 @@ defmodule EctoJob.Test.Repo.Migrations.AddJobTable do
   use Ecto.Migration
   alias EctoJob.Migrations.{CreateJobTable, Install}
 
+  @ecto_job_version 2
+
   def up do
     Install.up()
-    CreateJobTable.up("jobs")
+    CreateJobTable.up("jobs", version: @ecto_job_version)
   end
 
   def down do