+++ test/support/migrations/20191024105121_upgrade_job_table_v3.exs
@@ -1,0 +1,13 @@
+defmodule EctoJob.Test.Repo.Migrations.UpgradeJobTableV3 do
+  use Ecto.Migration
+
+  @ecto_job_version 3
+
+  def up do
+    EctoJob.Migrations.UpdateJobTable.up(@ecto_job_version, "jobs")
+  end
+
+  def down do
+    EctoJob.Migrations.UpdateJobTable.down(@ecto_job_version, "jobs")
+  end
+end